chore(batch-predictor,pyfunc-server): Bump merlin-sdk version in batch predictor and pyfunc server (#582)
# Description
This PR bumps up the maximum version of the `merlin-sdk` used in the
batch predictor and pyfunc server packages. This is a temporary measure
to allow the batch predictor and pyfunc server to use the latest version
of the SDK until the CICD pipelines get reworked to allow all 3
components to be published with the same version number.

# Modifications
- `python/batch-predictor/requirements.txt` - Bumping up of `merlin-sdk`
version
- `python/pyfunc-server/requirements.txt` - Bumping up of `merlin-sdk`
version

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE
```