Add pagination to the List Jobs UI (#587)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->
# Description
<!-- Briefly describe the motivation for the change. Please include
illustrations where appropriate. -->
This PR is based on the corresponding API PR
https://github.com/caraml-dev/merlin/pull/581 (once that is merged to
`main`, the base branch for this PR will be updated).

The per model version jobs list view now uses the new `/jobs-by-page`
paginated API. With this, the UI now enables pagination which is carried
out on the backend. Also, partial searching of the job name is possible
from the UI and this search is also carried out on the backend, using
the same API endpoint.

## Illustration


https://github.com/caraml-dev/merlin/assets/23465343/5a2ce290-3d45-4a10-b223-34d2e4e44816

# Modifications
<!-- Summarize the key code changes. -->
* `ui/src/config.js` - Introduce pagination page size config
* `ui/src/job/JobListTable.js` - Replace Eui in-memory table with a
basic table and implement the search bar and pagination
* `ui/src/job/Jobs.js` - Update backend API call

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->
- [x] Changes to search text resets pagination (back to page 1)

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
<!--
Does this PR introduce a user-facing change?
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
Add pagination to the List Jobs UI
```

---------

Co-authored-by: Arief Rahmansyah <ariefrahmansyah@hotmail.com>