chore(batch-predictor,pyfunc-server): Bump up merlin-sdk versions in batch predictor and pyfunc server (#597)
# Description
This PR pins the `merlin-sdk` versions of the batch predictor and pyfunc
server to the latest version available. (This is a temporary fix before
we modify the CICD workflow to publish new batch predictors and pyfunc
servers that are pinned to the latest `merlin-sdk` versions.)

# Modifications
- Pin `merlin-sdk` versions of the batch predictor and pyfunc server to
`==0.44.0`

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [ ] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE
```