feat(mlflow): Update Mlflow, API server, pyfunc server and batch predictor (#617)
# Description
This PR is the **second part** (2 out of 3) of a series of PRs to update
the version of Mlflow used by Merlin (to `1.26.1`*). These changes are:

1. Update the version of Mlflow used in the Merlin SDK and publish it to
PyPI
2. Update 
- the Merlin pyfunc server and batch predictor to use the updated Merlin
SDK version released in step 1 and publish them to PyPI
- the Merlin API server to ensure that requests sent to Mlflow reflect
the update API endpoint contracts
   - the Mlflow image so that it's built using version `1.26.1`
3. Update the default pyfunc server and batch predictor version in the
Merlin SDK and publish its new version to PyPI

# Main Modifications
- `.github/workflows/external.yml` - Updated the CICD job to release a
new version of the Mlflow image
- `api/mlflow/mlflow.go` - Added a header in all requests sent to the
Mlflow server from the Merlin API server because it's now required
- `api/mlflow/response.go` - Changed the data type of certain fields in
Mlflow's responses sent to Merlin
- `mlflow/Dockerfile` - Updated Dockerfile of the Mlflow server with
newer dependencies
- `python/batch-predictor/requirements.txt` - Updated the requirements
file with a newer version of the SDK which requires version `1.26.1` of
Mlflow
- `python/pyfunc-server/requirements.txt` - Updated the requirements
file with a newer version of the SDK which requires version `1.26.1` of
Mlflow

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
<!--
Does this PR introduce a user-facing change?
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```