feat(api,sdk): Add user-configurable secrets for deployments (#625)
# Description
This PR introduces a new feature that allows users to mount secrets
stored in MLP into their Merlin batch prediction workloads and Merlin
model deployments. This is similar to the existing feature whereby users
are able to mount their Google Cloud Platform's service account key
stored in MLP as a secret within the batch workloads.

## Batch jobs
The user-configured MLP secrets are basically stored within the
[existing secret that gets
created](https://github.com/caraml-dev/merlin/blob/967788eb98a0120dd20bec692ab04bcea8f08fc9/api/batch/controller.go#L166)
for each batch job run, before the Spark application gets created. The
Spark driver and executor access the user-configured MLP secrets as env
vars that have references to this k8s secret. This secret similarly gets
deleted after completion of the job.

## Model deployments
No existing process exists to create a k8s secret prior to the
deployment of a KServe inference service so this PR introduces some
changes to the deployment workflow to create this secret (not unlike
what happens for batch jobs). This secret gets deleted when the model is
undeployed, when there's a new successful deployment or when the
deployment fails. Similar to what happens for batch jobs, the KService
predictor (Merlin model) and transformer then access the user-configured
MLP secrets as env vars that have references to this k8s secret.

## DB migration
A small DB migration is required to create new columns in the
`version_endpoints` and `transformers` table to store the configured MLP
secrets. The batch jobs don't require this update as those secrets are
stored within an existing column that's already of the `jsonb` type.

This PR only introduces the necessary API server and SDK changes to make
this happen. The UI changes will be introduced in a separate PR to keep
this PR short.

# Modifications
- `api/batch/controller.go` - [Batch jobs] Minor refactoring to retrieve
GCP service account secret and user-configured secrets from MLP for
- `api/batch/resource.go` - [Batch jobs] Addition of env vars with
references to a k8s secret (containing all the user-configured secrets)
to the spark driver and executors
- `api/cluster/controller.go` - [Model deployments] Minor refactoring of
the existing cluster controller so that it now contains an MLP API
client object to retrieve MLP secrets with; introduction of additional
helper methods to create/delete secrets for predictors and transformers
- `api/cluster/resource/templater.go` - [Model deployments] Refactoring
of the KServe inference service manifest templater to add
user-configured secrets as env vars with references to a k8s secret
containing all the user-configured secrets)
- `api/cluster/secret.go` - [Model deployments] Addition of new k8s
secret creation/deletion methods using the cluster controller
- `api/cmd/api/main.go` - Minor refactoring to parse the MLP API client
to cluster controllers initialised
- `api/models/prediction_job.go` - [Batch jobs] Addition of the secrets
field to batch jobs
- `api/models/secret.go` - Addition of the new struct `Secret` that
contains information about the name of the secret as stored in MLP as
well as the name of the environment variable the secret will be
available as
- `api/models/service.go` - [Model deployments] Addition of the secrets
field to Merlin model services (this is just an intermediate struct used
by Merlin in the API server)
- `api/models/transformer.go` - [Model deployments] Addition of the
secrets field to Merlin transformers (this struct gets stored in the DB)
- `api/models/version_endpoint.go` - [Model deployments] Addition of the
secrets field to Merlin version endpoints (this struct gets stored in
the DB)
- `db-migrations/41_version_endpoints_add_secrets.**.sql` - Addition of
DB migration scripts to add a new column for secrets to be stored
- `python/sdk/merlin/endpoint.py` - Addition of steps to allow secrets
to be initialised correctly in a `VersionEndpoint` object
- `python/sdk/merlin/model.py` - Addition of steps to allow secrets to
be correctly parsed when the `deploy` function is called
- `python/sdk/merlin/mounted_mlp_secret.py` - Addition of a new class to
represent each user-configurable secret
- `swagger.yaml` - Update to swagger API specs

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [x] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE; DATABASE MIGRATION REQUIRED
```