feat(ui): Add secret management UI (#627)
# Description
Following up on the new feature allowing users to configure MLP secrets
in their model services/batch jobs as introduced in PR #625, this PR
introduces UI changes to complement that feature, allowing users to
configure secrets for model services and batch jobs, as well as to view
them in the details page view.

## Model Services
### Deployment Page
![Screenshot 2025-02-06 at 5 02
35 PM](https://github.com/user-attachments/assets/c386510d-cf94-4ea5-b303-80ba7a0c962b)

### Details Page
![Screenshot 2025-02-06 at 5 00
52 PM](https://github.com/user-attachments/assets/0af9d74e-a3b4-4ac0-b80f-2b8633f14cfc)

## Batch Jobs
### Submission Page
![Screenshot 2025-02-06 at 5 02
05 PM](https://github.com/user-attachments/assets/3dadcd4f-8c06-42a1-a797-240119bd1292)

### Details Page
![Screenshot 2025-02-06 at 5 01
03 PM](https://github.com/user-attachments/assets/0f85ba4a-afbb-4027-a5e5-b6284ded5ec1)

Note that the schema validation unfortunately only exists for model
service deployments and not batch jobs, and as such this PR does not add
additional new validation checks for the `secrets` field for batch jobs
triggered via the UI.

Some of the changes to the column widths/proportions were done to
improve consistency in the component sizes throughout the UI.

## Additional Change
This PR also includes a tiny fix to the API server to ensure that k8s
secrets for each Merlin transformer is only created when the transformer
is actually enabled (as opposed to when the transformer config exists).

# Modifications
- `api/cluster/controller.go` - Fix for the API server to not create
transformer secrets unnecessarily
- `ui/src/components/SecretsConfigTable.js` - New component to display
secrets in the details view
- `ui/src/pages/job/form/components/SecretsForm.js` - New form for users
to configure secrets for batch jobs
- `ui/src/pages/job/form/context.js` - Addition of a new context setter
for the secrets field
- `ui/src/pages/version/components/forms/components/SecretsPanel.js` -
New form for users to configure secrets for model services
- `ui/src/pages/version/components/forms/validation/schema.js` -
Addition of validation schema for secrets used in model services

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE
```