feat(api): Decouple Arize and BigQuery sink config (#630)
# Description
Use `BigQuery` sink as the sink default and add `EnabledModelSerials`
config to `ArizeSink` to determine which model needs `ArizeSink` for the
`ObservationSinks` config. The value for the `EnabledModelSerials` is
following this pattern `{Project}_{Model Name}_{Model Version}`,
separated with comma.

# Modifications
Only add `ArizeSink` in the consumer config if the model serial is in
the `EnabledModelSerials`

# Tests
1. Hit redeploy API locally while connected to staging env without
`OBSERVABILITYPUBLISHER_ARIZESINK_ENABLEDMODELSERIALS` env, the
`config.yaml` file generated for the consumer looks like this
```
observation_sinks:
- type: BIGQUERY
  config:
    project: p-gojek-data-science
    dataset: data_science_platform_playground
    ttl_days: 14
```
2. Hit redeploy API locally while connected to staging env with
`"OBSERVABILITYPUBLISHER_ARIZESINK_ENABLEDMODELSERIALS":
"tari-staging_pyfunc-play_3"` , the `config.yaml` file generated for the
consumer looks like this
```
observation_sinks:
- type: BIGQUERY
  config:
    project: p-gojek-data-science
    dataset: data_science_platform_playground
    ttl_days: 14
- type: ARIZE
  config:
    api_key: YzJjNmY5OGZhMGNmYTkxOTc0YQ==
    space_key: ODVkMmMwMg==
```

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

```release-note
NONE
```

---------

Co-authored-by: Muhammad Naufal Andika Natsir Putra <muhammad.putra@gojek.com>