chore(sdk): Bump up pyfunc package versions (#649)
# Description
Given that the current release of Merlin is `v0.48.3`, we need to update
the default version of the `merlin-batch-predictor` and
`merlin-pyfunc-server` Python packages.

## Additional Unrelated Changes 🚨
Due to the upgrading of `numpy` in this PR
[here](https://github.com/caraml-dev/merlin/pull/641#discussion_r2090771832),
Python 3.8 is no longer supported. The CICD pipeline fails because
Python 3.8 is currently still being used. This draft (at the time of
writing) PR #639 is supposed to include changes to remove support for
Python 3.8 but is temporarily paused due to other changes. @naufalandika

To fix the CICD pipeline for the moment, this PR copies the existing
changes in that PR to remove support entirely for Python 3.8 (and
includes other changes to other tests/examples that mention Python 3.8).

# Modifications
- `python/sdk/merlin/model.py` - Bumped up max version of the
`merlin-batch-predictor` and `merlin-pyfunc-server` Python packages

# Tests
<!-- Besides the existing / updated automated tests, what specific
scenarios should be tested? Consider the backward compatibility of the
changes, whether corner cases are covered, etc. Please describe the
tests and check the ones that have been completed. Eg:
- [x] Deploying new and existing standard models
- [ ] Deploying PyFunc models
-->

# Checklist
- [x] Added PR label
- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduces API
changes

# Release Notes
```release-note
NONE
```